[WAC-106] Fix bug introduced by ckanext-googleanalytics work
This commit fixes an issue that was introduced by WAC-34.
It does this by introducing a 'ckan_googleanalytics_enable' variable which when set to true enables the extension, not otherwise.
Enabling the extension means:
- adding the necessary configuration to 'production.ini'
- creating a secret containing the Google analytics service account secrets and mounting that to a file
- creating the cronjob that will fetch data from Google Analytics.

In addition, it also parametrises the 'googleanalytics.measurement_id' and 'googleanalytics.property_id' configuration
options so that setting up analytics is as easy as possible.